requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
groq>=1.1.0
sentence-transformers>=2.0.0
pypdf>=4.0.0
openpyxl>=3.1.0
numpy>=1.24.0
faiss-cpu>=1.7.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
itsdangerous>=2.1.0
xlrd>=2.0.0
authlib>=1.0.0
httpx>=0.24.0
